requirements.txt
absl-py==0.4.0
astor==0.7.1
awscli==1.16.7
azure==1.0.3
azure-common==1.1.14
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==2.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
backcall==0.1.0
better-exceptions==0.2.1
bleach==2.1.4
botocore==1.11.7
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.5.5
colorama==0.3.9
cycler==0.10.0
dask==0.19.0
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
entrypoints==0.2.3
fuel==0.2.0
future==0.16.0
gast==0.2.0
grpcio==1.14.1
gym==0.10.5
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imageio==2.3.0
ipdb==0.11
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
moviepy==0.2.3.5
mpi4py==3.0.0
mujoco-py==0.5.7
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numexpr==2.6.8
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
picklable-itertools==0.1.1
pickleshare==0.7.4
Pillow==5.2.0
progressbar2==3.38.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pyaml==17.12.1
pyasn1==0.4.4
pyglet==1.3.2
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tables==3.4.4
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
zmq==0.0.0
